+++ features/step_definitions/cli_steps.rb
@@ -96,6 +96,10 @@ When /^I modify the file "([^"]*)" to replace "([^"]*)" with "([^"]*)"$/ do |fil
   modify_file(file_name, orig_text, new_text)
 end
 
+When /^I append to file "([^"]*)":$/ do |file_name, content|
+  append_to_file(file_name, "\n" + content)
+end
+
 When /^I set the "([^"]*)" environment variable to "([^"]*)"$/ do |var, value|
   set_env(var, value)
 end
@@ -149,7 +153,8 @@ Then /^the file "([^"]*)" should contain a YAML fragment like:$/ do |file_name, 
 
     # Normalize by removing leading and trailing whitespace...
     file_content = file_content.split("\n").map do |line|
-      line.strip
+      # Different versions of psych use single vs. double quotes
+      line.strip.gsub('"', "'")
     end.join("\n")
 
     file_content.should include(fragment)