+++ features/support/env.rb
@@ -1,8 +1,8 @@
 require 'rubygems'
 require 'bundler'
 Bundler.setup
-require 'limited_red/plugins/cucumber'
 
+require 'limited_red/plugins/cucumber'
 require 'ruby-debug' if !defined?(RUBY_ENGINE) && RUBY_VERSION != '1.9.3' && !ENV['CI']
 
 require 'aruba/cucumber'