+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 [Full Changelog](http://github.com/myronmarston/vcr/compare/v2.0.0...master)
 
+* Add new `:use_scenario_name` option to the cucumber tags API. This
+  allows you to use a generic tag (such as `@vcr`) and have the
+  cassettes named based on the feature and scenario rather than based on
+  the tag. Thanks to [Omer Rauchwerger](https://github.com/rauchy) for
+  the implementation and [Chad Jolly](https://github.com/cjolly) for the
+  initial idea and feedback.
+* Add new `:decode_compressed_response` cassette option. When set to
+  true, VCR will decompress a gzipped or deflated response before
+  recording the cassette, in order to make it more human readable.
+  Thanks to [Mislav Marohnić](https://github.com/mislav) for the
+  idea and implementation.
+
 ## 2.0.0 (March 2, 2012)
 
 [Full Changelog](http://github.com/myronmarston/vcr/compare/v2.0.0.rc2...v2.0.0)